Finished new implementation of ResManager

+++ RUCengine/src/ResourceManager/ResPtr.h
@@ -18,30 +18,30 @@ namespace RUC
 		ResPtr(const ResPtr<S>& other)
 			:m_ResPointer((T*)other.m_ResPointer), m_Index(other.m_Index)
 		{
-			RUC_INFO("Old refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("Old refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 			ResourceManager::s_Metadata[m_Index].RefCount++;
-			RUC_INFO("New refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("New refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 		}
 
 		ResPtr(const ResPtr<T>& other)
 			:m_ResPointer(other.m_ResPointer), m_Index(other.m_Index)
 		{
-			RUC_INFO("Old refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("Old refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 			ResourceManager::s_Metadata[m_Index].RefCount++;
-			RUC_INFO("New refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("New refcount (Copy constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 		}
 
 		~ResPtr()
 		{
 			if (m_Index != -1)
 			{
-				RUC_INFO("Old refcount (Destructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+				//RUC_INFO("Old refcount (Destructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 				ResourceManager::s_Metadata[m_Index].RefCount--;
-				RUC_INFO("New refcount (Destructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+				//RUC_INFO("New refcount (Destructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 
-				if (ResourceManager::s_Metadata[m_Index].RefCount == 0)
+				if (ResourceManager::s_Metadata[m_Index].RefCount == 0 && !(ResourceManager::s_Metadata[m_Index].ResPtr->m_Flags & (char)ResourceFlags::PERSISTENT))
 				{
-					RUC_INFO("Object should be destroyed");
+					//RUC_INFO("Object should be destroyed");
 					ResourceManager::DeleteResource(m_Index);
 				}
 			}
@@ -63,9 +63,21 @@ namespace RUC
 			m_ResPointer = (T*)other.m_ResPointer;
 			m_Index = other.m_Index;
 
-			RUC_INFO("Old refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("Old refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 			ResourceManager::s_Metadata[m_Index].RefCount++;
-			RUC_INFO("New refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("New refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+
+			return *this;
+		}
+
+		ResPtr<T>& operator=(const ResPtr<T>& other)
+		{
+			m_ResPointer = other.m_ResPointer;
+			m_Index = other.m_Index;
+
+			//RUC_INFO("Old refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			ResourceManager::s_Metadata[m_Index].RefCount++;
+			//RUC_INFO("New refcount (Copy) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 
 			return *this;
 		}
@@ -87,9 +99,9 @@ namespace RUC
 		ResPtr(T* resource, int index)
 			: m_ResPointer(resource), m_Index(index)
 		{
-			RUC_INFO("Old refcount (Constructor of {1}) {0}", ResourceManager::s_Metadata[m_Index].RefCount, ResourceManager::s_Metadata[m_Index].ResPtr->GetName());
+			//RUC_INFO("Old refcount (Constructor of {1}) {0}", ResourceManager::s_Metadata[m_Index].RefCount, ResourceManager::s_Metadata[m_Index].ResPtr->GetName());
 			ResourceManager::s_Metadata[index].RefCount++;
-			RUC_INFO("New refcount (Constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
+			//RUC_INFO("New refcount (Constructor) {0}", ResourceManager::s_Metadata[m_Index].RefCount);
 		}
 
 	private:
@@ -101,80 +113,4 @@ namespace RUC
 
 		friend class ResourceManager;
 	};
-
-
-	/*template <typename T>
-	class ResPtr
-	{
-	public:
-		ResPtr(const ResPtr<T>& other)
-			:m_Resource(other.m_Resource)
-		{
-			//Increase refCount, a new pointer was created
-			m_Resource->m_RefCount++;
-		}
-
-		ResPtr<T>& operator=(const ResPtr<T>& other)
-		{
-			m_Resource = nullptr;
-
-			//Increase refCount, a new pointer was created
-			//m_Resource->m_RefCount++;
-
-			return *this;
-		}
-
-		ResPtr<Resource> getBasePointer()
-		{
-			return ResPtr((Resource*)m_Resource,)
-		}
-
-		~ResPtr()
-		{
-			int refCount = --m_Resource->m_RefCount;
-
-			if (refCount == 0)
-			{
-				delete m_Resource;
-			}
-		}
-
-		T* Get() const
-		{
-			return m_Resource;
-		}
-
-		//Cast a Resource ResPtr to a T ResPtr
-		static ResPtr<T> Cast(const ResPtr<Resource>& other)
-		{
-			return ResPtr<T>(dynamic_cast<T*>(other.Get()));
-		}
-
-	private:
-
-		ResPtr(Resource* resource, char flags)
-			: m_Resource(resource)
-		{
-			RUC_ASSERT(resource->m_RefCount == 0, "A pointer to the given resource already exists. Use a copy of that ResPtr instead of creating another one");
-
-			m_Resource->m_Flags = flags;
-		}
-		
-		ResPtr(T* resource)
-			: m_Resource(resource)
-		{
-		}
-
-		ResPtr()
-		{
-			m_Resource = nullptr;
-		}
-
-
-
-	private:
-		T* m_Resource;
-
-		friend class ResourceManager;
-	};*/
 }
